Revert "Fix/integrations"
