+++ prisma/schema.prisma
@@ -32,14 +32,6 @@ enum IntegrationConnectionStatus {
   DISCONNECTED
 }
 
-enum NotificationType {
-  ORDER_CREATED
-  MESSAGE_RECEIVED
-  TASK_CREATED
-  MEETING_CREATED
-  CALENDAR_EVENT_CREATED
-}
-
 model User {
   id                      String    @id @default(cuid())
   name                    String?
@@ -76,8 +68,6 @@ model User {
   orders            Order[]
   ownedDeals        Deal[]             @relation("DealOwner")
   automationRules   AutomationRule[]   @relation("AutomationOwner")
-
-  notifications     Notification[]
   assignedMessages  Message[]          @relation("MessageAssignee")
   internalComments  InternalComment[]  @relation("InternalCommentAuthor")
 }
@@ -164,24 +154,6 @@ model OAuthIntegration {
 
   @@unique([ownerId, provider])
   @@index([ownerId, status])
-}
-
-model Notification {
-  id        String           @id @default(cuid())
-  userId    String
-  type      NotificationType
-  title     String
-  body      String?          @db.Text
-  url       String?
-  entityId  String?
-  createdAt DateTime         @default(now())
-  readAt    DateTime?
-
-  user      User             @relation(fields: [userId], references: [id], onDelete: Cascade)
-
-  @@unique([userId, type, entityId])
-  @@index([userId, createdAt])
-  @@index([userId, readAt])
 }
 
 model VerificationToken {
