+++ src/app/api/webhooks/email/route.js
@@ -1,6 +1,5 @@
 import { NextResponse } from "next/server";
 import { prisma } from "@/lib/prisma";
-import { createNotification, NOTIFICATION_TYPES } from "@/lib/notifications";
 
 export async function POST(request) {
   try {
@@ -35,7 +34,7 @@ export async function POST(request) {
     const lead = hintedOwnerId
       ? await prisma.lead.findFirst({
           where: { ownerId: hintedOwnerId, email },
-          select: { id: true, ownerId: true },
+          select: { id: true },
         })
       : null;
 
@@ -47,7 +46,7 @@ export async function POST(request) {
         take: 2,
       });
       if (candidates.length === 1) {
-        resolvedLead = { id: candidates[0].id, ownerId: candidates[0].ownerId };
+        resolvedLead = { id: candidates[0].id };
       } else if (candidates.length > 1) {
         return NextResponse.json(
           {
@@ -64,7 +63,7 @@ export async function POST(request) {
       return NextResponse.json({ success: true, ignored: true }, { status: 200 });
     }
 
-    const message = await prisma.message.create({
+    await prisma.message.create({
       data: {
         leadId: resolvedLead.id,
         subject: String(subject),
@@ -73,17 +72,6 @@ export async function POST(request) {
         direction: "INBOUND",
       },
     });
-
-    if (resolvedLead.ownerId) {
-      await createNotification({
-        userId: resolvedLead.ownerId,
-        type: NOTIFICATION_TYPES.MESSAGE_RECEIVED,
-        title: "Nowa wiadomość",
-        body: String(subject),
-        url: `/dashboard/skrzynka?leadId=${encodeURIComponent(resolvedLead.id)}`,
-        entityId: message.id,
-      }).catch(() => null);
-    }
 
     return NextResponse.json({ success: true }, { status: 200 });
   } catch (e) {
